+++ demos/src/bin/platformer_3d.rs
@@ -4,9 +4,7 @@ use bevy::prelude::*;
 #[cfg(feature = "rapier3d")]
 use bevy_rapier3d::{prelude as rapier, prelude::*};
 use bevy_tnua::control_helpers::{
-    TnuaBlipReuseAvoidance, // TnuaCrouchEnforcer, TnuaCrouchEnforcerPlugin,
-    TnuaSimpleAirActionsCounter,
-    TnuaSimpleFallThroughPlatformsHelper,
+    TnuaBlipReuseAvoidance, TnuaSimpleAirActionsCounter, TnuaSimpleFallThroughPlatformsHelper,
 };
 #[allow(unused_imports)]
 use bevy_tnua::math::{AsF32, Vector3, float_consts};
@@ -93,14 +91,9 @@ fn main() {
         ScheduleToUse::Update => {
             // This is Tnua's main plugin.
             app.add_plugins(TnuaControllerPlugin::<DemoControlScheme>::new(Update));
-
-            // This plugin supports `TnuaCrouchEnforcer`, which prevents the character from standing up
-            // while obstructed by an obstacle.
-            // app.add_plugins(TnuaCrouchEnforcerPlugin::new(Update));
         }
         ScheduleToUse::FixedUpdate => {
             app.add_plugins(TnuaControllerPlugin::<DemoControlScheme>::new(FixedUpdate));
-            // app.add_plugins(TnuaCrouchEnforcerPlugin::new(FixedUpdate));
         }
     }
 
@@ -357,16 +350,6 @@ fn setup_player(
         );
         command_altering_selectors
     });
-
-    // `TnuaCrouchEnforcer` can be used to prevent the character from standing up when obstructed.
-    // cmd.insert(TnuaCrouchEnforcer::new(0.5 * Vector3::Y, |cmd| {
-    // #[cfg(feature = "rapier3d")]
-    // cmd.insert(TnuaRapier3dSensorShape(
-    // bevy_rapier3d::parry::shape::SharedShape::cylinder(0.0, 0.5),
-    // ));
-    // #[cfg(feature = "avian3d")]
-    // cmd.insert(TnuaAvian3dSensorShape(avian::Collider::cylinder(0.5, 0.0)));
-    // }));
 
     // The ghost sensor is used for detecting ghost platforms - platforms configured in the physics
     // backend to not contact with the character (or detect the contact but not apply physical