+++ demos/src/bin/shooter_like.rs
@@ -8,9 +8,7 @@ use bevy::window::{CursorGrabMode, CursorOptions, PrimaryWindow};
 use bevy_rapier3d::{prelude as rapier, prelude::*};
 // use bevy_tnua::builtins::TnuaBuiltinCrouch;
 use bevy_tnua::control_helpers::{
-    TnuaBlipReuseAvoidance, // TnuaCrouchEnforcer, TnuaCrouchEnforcerPlugin,
-    TnuaSimpleAirActionsCounter,
-    TnuaSimpleFallThroughPlatformsHelper,
+    TnuaBlipReuseAvoidance, TnuaSimpleAirActionsCounter, TnuaSimpleFallThroughPlatformsHelper,
 };
 use bevy_tnua::math::{AdjustPrecision, AsF32, Quaternion, float_consts};
 use bevy_tnua::{TnuaAnimatingState, TnuaGhostSensor, TnuaToggle};
@@ -94,14 +92,9 @@ fn main() {
         ScheduleToUse::Update => {
             // This is Tnua's main plugin.
             app.add_plugins(TnuaControllerPlugin::<DemoControlScheme>::new(Update));
-
-            // This plugin supports `TnuaCrouchEnforcer`, which prevents the character from standing up
-            // while obstructed by an obstacle.
-            // app.add_plugins(TnuaCrouchEnforcerPlugin::new(Update));
         }
         ScheduleToUse::FixedUpdate => {
             app.add_plugins(TnuaControllerPlugin::<DemoControlScheme>::new(FixedUpdate));
-            // app.add_plugins(TnuaCrouchEnforcerPlugin::new(FixedUpdate));
         }
     }
 
@@ -358,16 +351,6 @@ fn setup_player(
         );
         command_altering_selectors
     });
-
-    // `TnuaCrouchEnforcer` can be used to prevent the character from standing up when obstructed.
-    // cmd.insert(TnuaCrouchEnforcer::new(0.5 * Vector3::Y, |cmd| {
-    // #[cfg(feature = "rapier3d")]
-    // cmd.insert(TnuaRapier3dSensorShape(
-    // bevy_rapier3d::parry::shape::SharedShape::cylinder(0.0, 0.5),
-    // ));
-    // #[cfg(feature = "avian3d")]
-    // cmd.insert(TnuaAvian3dSensorShape(avian::Collider::cylinder(0.5, 0.0)));
-    // }));
 
     // The ghost sensor is used for detecting ghost platforms - platforms configured in the physics
     // backend to not contact with the character (or detect the contact but not apply physical