+++ src/builtins/walk_sensors.rs
@@ -5,8 +5,7 @@ use bevy_tnua_physics_integration_layer::data_for_backends::TnuaProximitySensor;
 #[derive(Copy, Clone)]
 pub struct TnuaBuiltinWalkSensors<'a> {
     pub ground: &'a TnuaProximitySensor,
-    // TODO:
-    //pub overhead: Option<&'a TnuaProximitySensor>,
+    pub headroom: Option<&'a TnuaProximitySensor>,
 }
 
 impl<'a> TnuaSensors<'a> for TnuaBuiltinWalkSensors<'a> {
@@ -16,6 +15,5 @@ impl<'a> TnuaSensors<'a> for TnuaBuiltinWalkSensors<'a> {
 #[derive(Default)]
 pub struct TnuaBuiltinWalkSensorsEntities {
     pub ground: Option<Entity>,
-    // TODO:
-    //pub overhead: Option<Entity>,
+    pub headroom: Option<Entity>,
 }